+++ ngbatis-demo/src/main/java/ye/weicheng/ngbatis/demo/repository/NoPropertiesVertexDao.java
@@ -1,0 +1,17 @@
+package ye.weicheng.ngbatis.demo.repository;
+
+// Copyright (c) 2022 All project authors. All rights reserved.
+//
+// This source code is licensed under Apache 2.0 License.
+
+import org.nebula.contrib.ngbatis.proxy.NebulaDaoBasic;
+import ye.weicheng.ngbatis.demo.pojo.NoPropertiesVertex;
+
+/**
+ * @author yeweicheng
+ * @since 2023-08-02 18:37
+ * <br>Now is history!
+ */
+public interface NoPropertiesVertexDao  extends NebulaDaoBasic<NoPropertiesVertex, String> {
+
+}